requirements.txt
Package             Version
------------------- -------
dataclasses         0.6    
Django              2.2.9  
djangorestframework 3.11.0 
djongo              1.2.38 
dnspython           1.16.0 
pip                 19.3.1 
pymongo             3.10.0 
pytz                2019.3 
setuptools          40.8.0 
sqlparse            0.2.4  
